Add Node.textContent polyfill for IE8
Using innerText as an "alias" of textContent is problematic, as there're
numerous discrepancies between the two properties. Most notably, reading
an element's innerText will skip any hidden properties (including script
or style tags), whereas textContent includes them.

Furthermore, relying on innerText to fill the role of textContent solves
only half the problem: it won't allow textContent to be used on comments
or text nodes. Trying to assign innerText on a non-Element in IE8 throws
a runtime error, so the following code would still break:

    // <div id="el">Text <b>Tag</b>  <!-- Comment --> </div>
    for(var i = 0; i < el.childNodes.length; ++i){
        el.childNodes[i].textContent = "Child #"+i;
    }

This polyfill uses a much more stable implementation, following the spec
as closely as possible (within IE8's limitations). Because of the way IE
collapses whitespace between elements, it's impossible for trailing line
breaks or indentation to be preserved. This is an issue with the way the
document tree is processed; otherwise, the behaviour of textContent will
be more-or-less consistent between IE8 and spec-compliant browsers.

More info:
  * http://perfectionkills.com/the-poor-misunderstood-innerText/
  * http://clubajax.org/plain-text-vs-innertext-vs-textcontent/